+++ NEWS.md
@@ -1,4 +1,10 @@
+# polyCheck 0.1.1
+
+* Added checks for self-intersection of the polygons, providing the ID of polygons which overlap in the diagnostics
+* Improved functionality of `display_polygon` to show settlement polygons
+
 # polyCheck 0.1.0
 
 * Initial release of the code
 * Added a `NEWS.md` file to track changes to the package.
+* Added basic diagnostics for single polygon